# Optimized outlier detection runtime