--- a/requirements.txt
+++ b/requirements.txt
@@ -5,3 +5,5 @@ scikit-learn==1.3.1
 scipy==1.9.3
 seaborn==0.11.1
 simple-settings==1.0.0
+datasets==2.14.6
+line-profiler==4.1.1