Optimized the outlier detection algorithm code in a few places for faster runtime. To test impact of changes to runtime, we first implemented:

- Runtime benchmark method: A script that evaluates the runtime of the outlier detection algorithm on the open-source Adult Census Income benchmarking dataset from the UCI ML repository. The method uses the timeit library to time the outlier detection code run on the data set 4 times. The script then takes the minimum of 5 of those iterations.
- Added line profiler functionality: Added "@profile" decorators to several of the algo's methods and utilized the line_profiler package to retrieve a line-by-line breakdown of runtime in the Python code.

For each algorithm modification, we ran the benchmarking code to measure the runtime of the algorithm. We compared this to the runtime before the modification to validate runtime improvement. We also ran the unit tests to ensure the algorithm still functioned as expected.

To target the next modification, we then ran the profiler and found the section of the code that had the largest share of the total run time. Code improvements were made through a combination of manual changes and consulting gpt-4 for the most efficient ways to rewrite sections of code. After changes were made, we validated the changes by again running the line profiler and benchmarking script. This process was repeated until changes were no longer resulting in meaningful run time improvements. Profiler is run using the following commands in the terminal:
`kernprof -l benchmark_outlierdetect.py`
`python -m line_profiler -rmt "benchmark_outlierdetect.py.lprof"`

We documented each code change and subsequent run time improvements below. Please note that this is only the time to run the outlier detection algorithm (separate from loading and preprocessing data) and that run time relative to the changes is the value to consider:

- Original algorithm code: 4.63 sec
- Moved the frequency summation inside the [MultinomialModel.compute_outlier_scores outside the first for loop ](https://github.com/dimagi/outlier-detect/blob/05c67157d00fc0c3f0ab8f6d1f1ef8f7a68287c4/outlierdetect/outlierdetect.py#L120): 2.27 sec
- In _run_alg method, changed the first for loop, which calculates [agg_col_to_data to use the pandas groupby functionality ](https://github.com/dimagi/outlier-detect/blob/05c67157d00fc0c3f0ab8f6d1f1ef8f7a68287c4/outlierdetect/outlierdetect.py#L324)(instead of iterating with comparisons): 1.70 sec

For final benchmarking with an observation of how number of aggregation units (i.e. workers) and questions impacts performance run time, we get the following output by running `python3 benchmark_outlierdetect.py `:

Benchmarking outlier detection algorithm

Observing impact of aggregation unit scale
    Aggregation unit scale = 0.01, Execution time: 1.62 seconds
    Aggregation unit scale = 0.02, Execution time: 3.09 seconds
    Aggregation unit scale = 0.03, Execution time: 4.59 seconds

Observing impact of number of questions
    Number of questions = 1, Execution time: 0.59 seconds
    Number of questions = 2, Execution time: 1.11 seconds
    Number of questions = 3, Execution time: 1.61 seconds