@bderenzi 2023-11-07
> I would like to remove the line_profiler decorators and functionality before merging into the main codebase for code cleanliness.

Agreed